timer controller built out and functional. Turned time into bigint

+++ db/migrate/20210210172916_create_timers.rb
@@ -1,7 +1,7 @@
 class CreateTimers < ActiveRecord::Migration[6.0]
   def change
     create_table :timers do |t|
-      t.integer :time
+      t.bigint :time
       t.string :title
       t.belongs_to :user, null: false, foreign_key: true
     end
